requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
docker==7.0.0
kubernetes==28.1.0
openai==1.3.0
anthropic==0.7.11
python-dateutil==2.8.2
PyYAML==6.0.1
Jinja2==3.1.2
click==8.1.7
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3
